feat: add multi-asset trade support (#148)
* feat(model): add multi-asset trade primitives

* feat(db): persist decimal quantities and settlement dates

* feat(core): support decimal multi-asset trade flows

* feat(brokers): map multi-asset order payloads

* feat(cli): expose multi-asset quantity handling

* fix(core): keep decimal sizing under risk caps

* test(cli): update decimal quantity JSON contracts

+++ core/src/calculators_trade/capital_not_at_risk.rs
@@ -50,7 +50,7 @@ impl TradeCapitalNotAtRisk {
                             trade.entry.unit_price, risk_per_share
                         )
                     })?
-                    .checked_mul(Decimal::from(trade.entry.quantity))
+                    .checked_mul(trade.entry.quantity)
                     .ok_or_else(|| {
                         format!(
                             "Arithmetic overflow in multiplication: {} * {}",
